Update handlebars and generated templates
Previously, the version of handlebars used was v1.3.0 whereas handlebars
is now at v3.0.2. This patch updates the handlebars runtime vendored
with the gem and includes regenerated templates since the compiled
templates APIs have changed a lot since v1.3.0. Upgrading to the latest
version of handlebars allows apps that are already using handlebars to
not to have to serve two different versions if they want to use this
handy gem.